+++ CHANGELOG.md
@@ -10,16 +10,19 @@
   * Added `ClothCollider` component
 * Added `AccelerationSmoothing` enum, defining gravity/winds acceleration smoothing
   * Added related `acceleration_smoothing` field to `ClothConfig`
+* Added custom anchor support with `VertexAnchor`
 
 ## API changes
 
 * (**BREAKING**) Renamed `ClothBuilder::fixed_points` to `anchored_vertex_ids`
   * Added `ClothBuilder::with_pinned_vertex_ids` method
+  * Added `ClothBuilder::with_pinned_vertex_id` method
   * Added `ClothBuilder::with_anchored_vertex_ids` method
   * Added `ClothBuilder::with_anchored_vertex_id` method
   * Deprecated `ClothBuilder::with_fixed_points` in favor of `ClothBuilder::with_pinned_vertex_ids`
 * Added `ClothBuilder::anchored_vertex_colors` field:
   * Added `ClothBuilder::with_pinned_vertex_colors` method
+  * Added `ClothBuilder::with_pinned_vertex_color` method
   * Added `ClothBuilder::with_anchored_vertex_colors` method
   * Added `ClothBuilder::with_anchored_vertex_color` method
 * Added `ClothBuilder::with_flat_normals` method